+++ src/headers/ui/src/ModItem.cpp
@@ -24,6 +24,7 @@ bool ModItem::init(Mod* mod, CCSize const& size, FavoritesPopup* parentPopup, bo
 
         // Background for mod item
         m_backgroundSprite = CCScale9Sprite::create("square02b_001.png");
+        m_backgroundSprite->setID("background");
         m_backgroundSprite->setScale(0.5f);
         m_backgroundSprite->setContentSize({ size.width * 2.f, size.height * 2.f });
         m_backgroundSprite->setAnchorPoint({ 0.5, 0.5 });
@@ -36,6 +37,7 @@ bool ModItem::init(Mod* mod, CCSize const& size, FavoritesPopup* parentPopup, bo
 
         // Mod icon sprite
         auto modIcon = geode::createModLogo(m_mod->getPackagePath());
+        modIcon->setID("mod-icon");
         modIcon->setScale(0.5f);
         modIcon->setPosition({ 20.f, getScaledContentHeight() / 2.f });
         modIcon->ignoreAnchorPointForPosition(false);
@@ -45,27 +47,43 @@ bool ModItem::init(Mod* mod, CCSize const& size, FavoritesPopup* parentPopup, bo
 
         // Mod name label
         auto nameLabel = CCLabelBMFont::create(m_mod->getName().c_str(), "bigFont.fnt");
+        nameLabel->setID("mod-name");
         nameLabel->setPosition({ 37.5f, 25.f });
         nameLabel->setScale(0.4f);
-        nameLabel->setAnchorPoint({ 0.f, 0.5f });
+        nameLabel->setAnchorPoint({ 0, 0.5 });
 
         addChild(nameLabel);
 
+        auto devs = m_mod->getDevelopers();
+        auto devLabelText = devs[0];
+        int andMore = as<int>(devs.size()) - 1;
+
+        if (andMore > 0) devLabelText += " & " + std::to_string(andMore) + " more";
+
+        // Mod name label
+        auto devLabel = CCLabelBMFont::create(devLabelText.c_str(), "goldFont.fnt");
+        devLabel->setID("mod-developers");
+        devLabel->setPosition({ nameLabel->getScaledContentWidth() + 40.f, 25.f });
+        devLabel->setScale(0.25f);
+        devLabel->setAnchorPoint({ 0, 0.5 });
+
+        addChild(devLabel);
+
         // Version label
         auto versionLabel = CCLabelBMFont::create(m_mod->getVersion().toVString().c_str(), "goldFont.fnt");
+        versionLabel->setID("mod-version");
         versionLabel->setPosition({ 37.5f, 15.f });
         versionLabel->setScale(0.3f);
-        versionLabel->setAnchorPoint({ 0.f, 0.5f });
-        versionLabel->setColor({ 200, 200, 200 });
+        versionLabel->setAnchorPoint({ 0, 0.5 });
 
         addChild(versionLabel);
 
         // ID label
         auto idLabel = CCLabelBMFont::create(m_mod->getID().c_str(), "bigFont.fnt");
+        idLabel->setID("mod-id");
         idLabel->setPosition({ versionLabel->getScaledContentWidth() + 40.f, 15.f });
         idLabel->setScale(0.25f);
-        idLabel->setAnchorPoint({ 0.f, 0.5f });
-        idLabel->setColor({ 125, 125, 125 });
+        idLabel->setAnchorPoint({ 0, 0.5 });
         idLabel->setOpacity(125);
 
         addChild(idLabel);
@@ -79,6 +97,7 @@ bool ModItem::init(Mod* mod, CCSize const& size, FavoritesPopup* parentPopup, bo
             this,
             menu_selector(ModItem::onViewMod)
         );
+        viewBtn->setID("view-button");
 
         // Favorite button here :)
         auto favBtnSprite = CCSprite::createWithSpriteFrameName(isFavorite ? "GJ_starsIcon_001.png" : "GJ_starsIcon_gray_001.png");
@@ -89,22 +108,24 @@ bool ModItem::init(Mod* mod, CCSize const& size, FavoritesPopup* parentPopup, bo
             this,
             menu_selector(ModItem::onFavorite)
         );
+        m_favButton->setID("favorite");
 
-        // Create menu for buttons
-        auto btnMenu = CCMenu::create();
-        btnMenu->setAnchorPoint({ 1, 0.5 });
-        btnMenu->setPosition({ getScaledContentWidth() - 10.f, getScaledContentHeight() / 2.f });
-        btnMenu->setScaledContentSize({ getScaledContentWidth() * 0.375f, getScaledContentHeight() - 10.f });
-
-        // Layout to automatically position buttons on menu
+        // Layout to automatically position buttons on button menu
         auto btnMenuLayout = RowLayout::create();
         btnMenuLayout->setDefaultScaleLimits(0.625f, 0.875f);
         btnMenuLayout->setAxisAlignment(AxisAlignment::End);
         btnMenuLayout->setCrossAxisAlignment(AxisAlignment::Center);
         btnMenuLayout->setCrossAxisLineAlignment(AxisAlignment::Center);
+        btnMenuLayout->setAutoGrowAxis(true);
         btnMenuLayout->setAxisReverse(true);
         btnMenuLayout->setGap(7.5f);
 
+        // Create menu for buttons
+        auto btnMenu = CCMenu::create();
+        btnMenu->setID("button-menu");
+        btnMenu->setAnchorPoint({ 1, 0.5 });
+        btnMenu->setPosition({ getScaledContentWidth() - 10.f, getScaledContentHeight() / 2.f });
+        btnMenu->setScaledContentSize({ getScaledContentWidth() * 0.375f, getScaledContentHeight() - 10.f });
         btnMenu->setLayout(btnMenuLayout);
 
         btnMenu->addChild(viewBtn);
@@ -113,6 +134,15 @@ bool ModItem::init(Mod* mod, CCSize const& size, FavoritesPopup* parentPopup, bo
         addChild(btnMenu);
 
         btnMenu->updateLayout();
+
+        if (auto helpTxt = firstTimeText()) {
+            log::debug("Adding help text...");
+
+            helpTxt->setAnchorPoint({ 1, 0.5 });
+            helpTxt->setPosition({ getScaledContentWidth() - (btnMenu->getScaledContentWidth() + 2.5f), getScaledContentHeight() / 2.f });
+
+            addChild(helpTxt);
+        };
 
         return true;
     } else {
@@ -140,15 +170,32 @@ void ModItem::onFavorite(CCObject*) {
 
 void ModItem::updateFavoriteIcon() {
     if (m_favButton) {
-        auto newSprite = CCSprite::createWithSpriteFrameName(
-            isFavorite ? "GJ_starsIcon_001.png" : "GJ_starsIcon_gray_001.png"
-        );
+        auto newSprite = CCSprite::createWithSpriteFrameName(isFavorite ? "GJ_starsIcon_001.png" : "GJ_starsIcon_gray_001.png");
         newSprite->setScale(0.875f);
 
         m_favButton->setNormalImage(newSprite);
     } else {
         log::error("Favorite button not found");
     };
+};
+
+CCLabelBMFont* ModItem::firstTimeText() {
+    if (m_thisMod->getSavedValue<bool>("already-loaded") || m_thisMod->hasSavedValue(m_thisMod->getID())) { // check if mod loaded before
+        return nullptr;
+    } else if (m_mod->getID().compare(m_thisMod->getID()) == 0) { // create the help text if loaded for the first time
+        log::info("Mod loaded for the first time, creating help text...");
+
+        // Help text for first-time users
+        auto help = CCLabelBMFont::create("Press to toggle ->", "chatFont.fnt");
+        help->setID("first-time-help-text");
+        help->setScale(0.5f);
+        help->setColor({ 200, 200, 200 });
+        help->setOpacity(200);
+
+        return help;
+    };
+
+    return nullptr;
 };
 
 ModItem* ModItem::create(Mod* mod, CCSize const& size, FavoritesPopup* parentPopup, bool geodeTheme) {