+++ changelog.md
@@ -1,3 +1,10 @@
+# v1.0.3
+- Added developer label to mod list item
+- Fixed coloring for version and ID label in mod list item
+- Adjusted positioning for clear search button and search box on pop-up
+- Added guide text for first-time users on *Favorite Mods* mod list item
+- Added and tweaked more node IDs
+
 # v1.0.2
 - Fixed touch priority issue with scrolling UI
 - Fixed texture size issue with Clear button